move graphql into services in next-client
