+++ apps/next-client/src/app/[locale]/dato-cms/client-fetch/layout.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { ClientContext } from 'graphql-hooks';
-import { datoCmsClient } from '@/graphql/client';
+import { datoCmsClient } from '@/services/graphql';
 
 type LayoutProps = {
   children: React.ReactNode;
