+++ apps/next-client/src/services/graphql/index.ts
@@ -1,0 +1,2 @@
+export * from './client';
+export * from './queries';
