+++ apps/next-client/src/services/graphql/queries.ts
