--- apps/next-client/src/services/index.ts
@@ -1,1 +1,0 @@
-export { default as mongoDBService } from './mongoDB';
